Add missing hash from store
